Several fixes for KvNomad
Here I fixed the following problems in the light theme (didn't have time for the dark one but you might want to make some of these changes to it):

(1) Similar *complex* elements are cloned (linked) insteaad of drawn again. In this way, the SVG file size is reduced, which is important because the lighter is the SVG file, the less is the memory usage.

(2) When the pressed, focused or toggled state is identical to the normal state, it could be removed (→ lighter SVG).

(3) Some complex SVG elements had redundant rectangles, some of which are removed (→ lighter SVG).

(4) Item views were to high. They're made tighter and their SVG images are modified accordingly. Also inactive SVG elelments are added.

(5) Since in this theme, line-edits and combos are drawn like buttons, they could use the button element. So, their separate elements are removed from the SVG file and their frame and interior are inherited from `PanelButtonCommand` now (→ lighter SVG).

(6) Kvantum reduces the frame size to 3px and removes frame expansion if an app (like KCalc) doesn't give enough size to buttons, combos or line-edits (bad app design). If the current button elements were used for non-expanded frames, their borders might be invisible. So, `frame.expandedElement` is set to `button` and `frame.element` (non-expanded frame) is set to `nbutton`. `nbutton` elements are added to the SVG file so that their frames is a multiplier of 3 instead of 6. You may compare the look of KCalc buttons, for example, before and after these changes.

(5) Inactive highlight color is corrected.

(6) Header and scrollbar needed different arrows.

And...

The SVG file can still be made lighter by more cloning (of arrows) and there may be other minor issues but I didn't have more time. You might also want to run kavntumpreview in the terminal to know about probable issues in the SVG file (there wil be warnings if there's an issue).

Last but not least, this theme is elegant. Please fix the issues of its dark version too if there are any.